% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_generator.R
\name{generate_fisheries_data}
\alias{generate_fisheries_data}
\title{Generate dummy fisheries data}
\usage{
generate_fisheries_data()
}
\value{
A data frame with fisheries data
}
\description{
Generate dummy fisheries data
}
