% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_generator.R
\name{generate_vessel_data}
\alias{generate_vessel_data}
\title{Generate vessel information}
\usage{
generate_vessel_data()
}
\value{
A data frame with vessel details
}
\description{
Generate vessel information
}
